+++ src/kafka/types/ITransactionContext.ts
@@ -3,6 +3,9 @@ export interface ITransactionContext {
   to?: string
   gas?: number
   gasPrice?: string
+  gasFeeCap?: string
+  gasTipCap?: string
+  txType?: string
   nonce?: number
   value?: string
   input?: string