+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## 5.0.0 (WIP)
+
+### 🆕 Features
+
+- Support for new transaction pricing mechanism (eip-1559)
+
+### ⚠ BREAKING CHANGES
+
+- Removed `/accounts/{address}/sign` and `/accounts/{address}/verify-signature` in favor of `/accounts/{address}/sign-message` and `/accounts/verify-message` accordingly to EIP-191 standards
+
 ## 4.2.1 (2021-10-06)
 
 ### 🛠 Bug fixes